refactor(pipeline): extract inference text helpers
